feat: Implement entry and theme converters, and update API handler interfaces
